% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacProfileInfo.R
\name{cacProfileInfo}
\alias{cacProfileInfo}
\title{CAC profile information}
\usage{
cacProfileInfo(cacFit, eta = 1e-04)
}
\arguments{
\item{cacFit}{An S3 object of class 'cacFit' returned from the cac function.}

\item{eta}{Determines the step size around the MLE for numerical differentiation.}
}
\value{
A list containing the following:
\describe{
  \item{sePi}{Vector of standard errors corresponding to the mixture probability estimates.}
  \item{seBeta}{Matrix of standard errors corresponding to the coefficient estimates.}
  \item{seInt}{Vector of standard errors corresponding to the intercept estimates, which
  represent the log hazard ratio between the baseline hazard functions. Only applicable
  when \code{hazRestrict = "proportional"}.}
  \item{var}{Variance matrix estimate for pi, intercepts (if applicable), and beta,
  in that order.}
  \item{profileInfo}{The approximate observed information matrix. Has fewer components
  than the variance matrix because baseline classes for pi and intercept (if applicable)
  estimators are not included, because these would make the matrix non-invertible.}
  \item{profileInfoRaw}{The approximate observed information matrix, before adjusting
  for symmetry. \code{profileInfo = (profileInfo + t(profileInfo)) / 2}.}
}
Each item has NDM and NDS versions, corresponding to the two approximation methods.
}
\description{
Calculates standard errors for pi and beta (and intercepts under the proportional
baseline hazard model). The observed information of the semiparametric likelihood
is approximated by two numerical differentiation techniques: NDM and NDS.
First order Richardson extrapolation is used for both techniques.
The NDM approach differentiates the EM algorithm operator, and the NDS approach
numerically differentiates the profile score function. These methods should return
similar approximations.
}
